refactor: controller 내부 패키지 정리
