+++ src/main/java/lotto/controller/view/Regrex.java
@@ -1,4 +1,4 @@
-package lotto.controller;
+package lotto.controller.view;
 
 public enum Regrex {
